implementing code best practices
